#25 Refactor: 헥사고날 아키텍쳐에 맞게 패키지 구조를 정리한다
