+++ src/main/java/hanieum/conik/domain/company/Company.java
@@ -1,6 +1,6 @@
 package hanieum.conik.domain.company;
 
-import hanieum.conik.domain.proposal.domain.entity.AbstractEntity;
+import hanieum.conik.global.domain.AbstractEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
