+++ src/main/java/hanieum/conik/domain/project/entity/ProjectRequest.java
@@ -1,6 +1,6 @@
 package hanieum.conik.domain.project.entity;
 
-import hanieum.conik.domain.proposal.domain.entity.AbstractEntity;
+import hanieum.conik.global.domain.AbstractEntity;
 import jakarta.persistence.Entity;
 import lombok.AccessLevel;
 import lombok.Getter;
